+++ test/test-worker.js
@@ -1,41 +1,16 @@
-const {
-	Worker, MessageChannel, MessagePort, isMainThread, parentPort
-} = require('worker_threads');
+require('../src');
+require('../src/threads/threadManager');
 
-const Tester1 = () => {
-	if (isMainThread) {
-		const worker = new Worker(__filename);
-		const subChannel = new MessageChannel();
-		worker.postMessage({ hereIsYourPort: subChannel.port1, message: 'What The Fuck 1...' }, [subChannel.port1]);
-		subChannel.port2.on('message', (value) => {
-			console.log('received:', value);
-		});
-	}
-	else {
-		parentPort.once('message', (value) => {
-			console.log('msg worker 1 actived');
-			value.hereIsYourPort.postMessage('the worker 1 is sending "' + value.message + '"');
-			value.hereIsYourPort.close();
-		});
-	}
-};
+const ThreadManager = _('Utils.Threads');
 
-const Tester2 = () => {
-	if (isMainThread) {
-		const worker = new Worker(__filename);
-		worker.postMessage({ message: 'What The Fuck 2...' });
-		worker.on('message', value => {
-			console.log('received:', value);
-		});
-	}
-	else {
-		parentPort.once('message', value => {
-			console.log('msg worker 2 actived');
-			parentPort.postMessage('the worker 2 is sending "' + value.message + '"');
-			parentPort.close();
-		});
-	}
-};
+const worker = ThreadManager.create('~/test/worker1.js', {});
 
-// Tester1();
-Tester2();
+(async () => {
+	console.log('step 1');
+	worker.send('Aloha Kosmos!');
+	console.log('step 2');
+	worker.request('fuck', 'you');
+	console.log('step 3');
+	await worker.requestAndWait('fuck', 'you all');
+	console.log('step 4');
+})();